requirements.txt
alembic==1.13.2
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
reportlab==4.2.2
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
